Update hide/show salary break down and add count down timer
